Add files via upload
This  is the basic view to company website (DIGISOLVE). This is been made using MEDIA QUERY jus for the ease of access of the website in the different media according to the width and the size of the screen. HTML- the basic code and the tag ,elements are used and the links to connect the diiferent files to the body. CSS is being used for the styling purpose with different selectors and blocks.